Remove back button from web
