--- packages/article-list/src/styles/index.android.js
@@ -1,8 +1,0 @@
-import { StyleSheet } from "react-native";
-import sharedStyles from "./shared.native";
-
-const styles = StyleSheet.create({
-  ...sharedStyles
-});
-
-export default styles;
